refactor(options): split OptionLoaderService into collaborators
Phase B of #365: product value loading, admin fields, and caption
overlays become separate classes; batch prefetch stays on the product
loader. Facade keeps the public API used by OptionService and mgr.
